chore: add intermediary server component to reduce page duplication
